+++ Demos/Python/DemoADMM_2D.py
@@ -75,7 +75,7 @@ RectoolsDIR = RecToolsDIR(DetectorsDimH = P,  # DetectorsDimH # detector dimensi
                     CenterRotOffset = None, # Center of Rotation (CoR) scalar (for 3D case only)
                     AnglesVec = angles_rad, # array of angles in radians
                     ObjSize = N_size, # a scalar to define reconstructed object dimensions
-                    device='gpu')
+                    device_projector='gpu')
 
 FBPrec = RectoolsDIR.FBP(noisy_sino) #perform FBP
 
@@ -97,19 +97,23 @@ Rectools = RecToolsIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension (
                     AnglesVec = angles_rad, # array of angles in radians
                     ObjSize = N_size, # a scalar to define reconstructed object dimensions
                     datafidelity='LS',# data fidelity, choose LS, PWLS (wip), GH (wip), Student (wip)
-                    nonnegativity='ENABLE', # enable nonnegativity constraint (set to 'ENABLE')
                     OS_number = None, # the number of subsets, NONE/(or > 1) ~ classical / ordered subsets
-                    tolerance = 1e-06, # tolerance to stop OUTER iterations earlier
-                    device='gpu')
+                    device_projector='gpu')
 
+# prepare dictionaries with parameters:
+data = {'projection_norm_data' : noisy_sino} # data dictionary
+algorithm_params = {'iterations' : 15,
+                    'ADMM_rho_const' : 4000.0
+                    }
+
+# adding regularisation using the CCPi regularisation toolkit
+regularisation_params = {'method' : 'FGP_TV',
+                         'regul_param' : 0.06,
+                         'iterations' : 100,
+                         'device_regulariser': 'gpu'}
 
 # Run ADMM reconstrucion algorithm with regularisation
-RecADMM_reg = Rectools.ADMM(noisy_sino,
-                              rho_const = 4000.0, \
-                              iterationsADMM = 15, \
-                              regularisation = 'SB_TV', \
-                              regularisation_parameter = 0.06,\
-                              regularisation_iterations = 50)
+RecADMM_reg = Rectools.ADMM(data, algorithm_params, regularisation_params)
 
 plt.figure()
 plt.imshow(RecADMM_reg, vmin=0, vmax=1, cmap="gray")