+++ Demos/Python/DemoFISTA_artifacts2D.py
@@ -78,7 +78,7 @@ Rectools = RecToolsDIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension 
                     CenterRotOffset = None, # Center of Rotation (CoR) scalar (for 3D case only)
                     AnglesVec = angles_rad, # array of angles in radians
                     ObjSize = P, # a scalar to define reconstructed object dimensions
-                    device='gpu')
+                    device_projector='gpu')
 
 print ("%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%")
 print ("Reconstructing analytical sinogram using FBP (tomobar)...")
@@ -89,32 +89,23 @@ FBPrec_misalign = Rectools.FBP(noisy_sino_misalign) # reconstruction with misali
 
 plt.figure()
 plt.subplot(131)
-plt.imshow(FBPrec_ideal, vmin=0, vmax=3, cmap="gray")
-plt.colorbar(ticks=[0, 0.5, 3], orientation='vertical')
+plt.imshow(FBPrec_ideal, vmin=0, vmax=2, cmap="gray")
+plt.colorbar(ticks=[0, 0.5, 2], orientation='vertical')
 plt.title('Ideal FBP reconstruction')
 plt.subplot(132)
-plt.imshow(FBPrec_error, vmin=0, vmax=3, cmap="gray")
-plt.colorbar(ticks=[0, 0.5, 3], orientation='vertical')
+plt.imshow(FBPrec_error, vmin=0, vmax=2, cmap="gray")
+plt.colorbar(ticks=[0, 0.5, 2], orientation='vertical')
 plt.title('Erroneous data FBP Reconstruction')
 plt.subplot(133)
 plt.imshow(FBPrec_misalign, vmin=0, vmax=3, cmap="gray")
-plt.colorbar(ticks=[0, 0.5, 3], orientation='vertical')
+plt.colorbar(ticks=[0, 0.5, 2], orientation='vertical')
 plt.title('Misaligned noisy FBP Reconstruction')
 plt.show()
 
 plt.figure() 
-plt.imshow(abs(FBPrec_ideal-FBPrec_error), vmin=0, vmax=3, cmap="gray")
-plt.colorbar(ticks=[0, 0.5, 3], orientation='vertical')
+plt.imshow(abs(FBPrec_ideal-FBPrec_error), vmin=0, vmax=2, cmap="gray")
+plt.colorbar(ticks=[0, 0.5, 2], orientation='vertical')
 plt.title('FBP reconsrtuction differences')
-#%%
-# %TESTING
-forwproj = Rectools.FORWPROJ(FBPrec_error)
-residual = noisy_zing_stripe - forwproj
-
-from tomobar.supp.addmodules import RING_WEIGHTS
-
-ring_weights2 = RING_WEIGHTS(residual,  window_halfsize_detectors = 9, window_halfsize_angles = 5, window_halfsize_projections = 0)
-
 #%%
 print ("%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%")
 print ("Reconstructing using FISTA method (tomobar)")
@@ -125,43 +116,29 @@ RectoolsIR = RecToolsIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension
                     CenterRotOffset = None, # Center of Rotation (CoR) scalar (for 3D case only)
                     AnglesVec = angles_rad, # array of angles in radians
                     ObjSize = P, # a scalar to define reconstructed object dimensions
-                    datafidelity='LS', #data fidelity, choose LS, PWLS, GH (wip), Student (wip)
-                    nonnegativity='ENABLE', # enable nonnegativity constraint (set to 'ENABLE')
+                    datafidelity='LS', #data fidelity, choose LS
                     OS_number = None, # the number of subsets, NONE/(or > 1) ~ classical / ordered subsets
-                    tolerance = 1e-06, # tolerance to stop outer iterations earlier
-                    device='gpu')
+                    device_projector='gpu')
 
-lc = RectoolsIR.powermethod() # calculate Lipschitz constant
+# prepare dictionaries with parameters:
+data = {'projection_norm_data' : noisy_zing_stripe} # data dictionary
+lc = RectoolsIR.powermethod(data) # calculate Lipschitz constant (run once to initialise)
+algorithm_params = {'iterations' : 350,
+                    'lipschitz_const' : lc}
 
+# adding regularisation using the CCPi regularisation toolkit
+regularisation_params = {'method' : 'FGP_TV',
+                         'regul_param' : 0.001,
+                         'iterations' : 250,
+                         'device_regulariser': 'gpu'}
 
-# Run FISTA reconstrucion algorithm with regularisation 
-RecFISTA_LS_reg = RectoolsIR.FISTA(noisy_zing_stripe,
-                                   iterationsFISTA = 350, 
-                                   regularisation = 'ROF_TV', 
-                                   regularisation_parameter = 0.003,
-                                   lipschitz_const = lc)
-plt.figure()
-plt.imshow(RecFISTA_LS_reg, vmin=0, vmax=3, cmap="gray")
-plt.colorbar(ticks=[0, 0.5, 3], orientation='vertical')
-plt.title('FISTA-LS-TV reconstruction')
-#%%
-RectoolsIR = RecToolsIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension (horizontal)
-                    DetectorsDimV = None,  # DetectorsDimV # detector dimension (vertical) for 3D case only
-                    CenterRotOffset = None, # Center of Rotation (CoR) scalar (for 3D case only)
-                    AnglesVec = angles_rad, # array of angles in radians
-                    ObjSize = P, # a scalar to define reconstructed object dimensions
-                    datafidelity='LS', #data fidelity, choose LS, PWLS, GH (wip), Student (wip)
-                    nonnegativity='ENABLE', # enable nonnegativity constraint (set to 'ENABLE')
-                    OS_number = None, # the number of subsets, NONE/(or > 1) ~ classical / ordered subsets
-                    tolerance = 1e-06, # tolerance to stop outer iterations earlier
-                    device='gpu')
+# Run FISTA reconstrucion algorithm with regularisation
+RecFISTA_LS_reg = RectoolsIR.FISTA(data, algorithm_params, regularisation_params)
 
-RecFISTA_Huber_reg = RectoolsIR.FISTA(noisy_zing_stripe, 
-                                   huber_data_threshold = 3.5,
-                                   iterationsFISTA = 350, 
-                                   regularisation = 'ROF_TV', 
-                                   regularisation_parameter = 0.003,
-                                   lipschitz_const = lc)
+# adding Huber data fidelity threshold 
+data.update({'huber_threshold' : 4.5})
+# Run FISTA reconstrucion algorithm with regularisation and HUber data
+RecFISTA_Huber_reg = RectoolsIR.FISTA(data, algorithm_params, regularisation_params)
 
 plt.figure()
 plt.subplot(121)
@@ -191,170 +168,93 @@ RectoolsIR = RecToolsIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension
                     CenterRotOffset = None, # Center of Rotation (CoR) scalar (for 3D case only)
                     AnglesVec = angles_rad, # array of angles in radians
                     ObjSize = P, # a scalar to define reconstructed object dimensions
-                    datafidelity='LS', #data fidelity, choose LS, PWLS, Huber, GH (wip), Student (wip)
-                    nonnegativity='ENABLE', # enable nonnegativity constraint (set to 'ENABLE')
-                    OS_number = 12, # the number of subsets, NONE/(or > 1) ~ classical / ordered subsets
-                    tolerance = 1e-06, # tolerance to stop outer iterations earlier
-                    device='gpu')
+                    datafidelity='LS', #data fidelity, choose LS
+                    OS_number = 8, # the number of subsets, NONE/(or > 1) ~ classical / ordered subsets
+                    device_projector='gpu')
 
 
-lc = RectoolsIR.powermethod() # calculate Lipschitz constant
+# prepare dictionaries with parameters:
+data = {'projection_norm_data' : noisy_zing_stripe} # data dictionary
+lc = RectoolsIR.powermethod(data) # calculate Lipschitz constant (run once to initialise)
+algorithm_params = {'iterations' : 20,
+                    'lipschitz_const' : lc}
+
+# adding regularisation using the CCPi regularisation toolkit
+regularisation_params = {'method' : 'FGP_TV',
+                         'regul_param' : 0.001,
+                         'iterations' : 100,
+                         'device_regulariser': 'gpu'}
+
 
 # Run FISTA reconstrucion algorithm with regularisation 
-RecFISTA_LS_reg = RectoolsIR.FISTA(noisy_zing_stripe,
-                                   iterationsFISTA = 20, 
-                                   regularisation = 'ROF_TV', 
-                                   regularisation_parameter = 0.01,
-                                   regularisation_iterations = 300,
-                                   lipschitz_const = lc)
-#%%
-RectoolsIR = RecToolsIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension (horizontal)
-                    DetectorsDimV = None,  # DetectorsDimV # detector dimension (vertical) for 3D case only
-                    CenterRotOffset = None, # Center of Rotation (CoR) scalar (for 3D case only)
-                    AnglesVec = angles_rad, # array of angles in radians
-                    ObjSize = P, # a scalar to define reconstructed object dimensions
-                    datafidelity='LS', #data fidelity, choose LS, PWLS, GH (wip), Student (wip)
-                    nonnegativity='ENABLE', # enable nonnegativity constraint (set to 'ENABLE')
-                    OS_number = 12, # the number of subsets, NONE/(or > 1) ~ classical / ordered subsets
-                    tolerance = 1e-06, # tolerance to stop outer iterations earlier
-                    device='gpu')
+RecFISTA_LS_reg = RectoolsIR.FISTA(data, algorithm_params, regularisation_params)
 
-RecFISTA_Huber_reg = RectoolsIR.FISTA(noisy_zing_stripe, 
-                                   huber_data_threshold=3.0,
-                                   iterationsFISTA = 20, 
-                                   regularisation = 'ROF_TV', 
-                                   regularisation_parameter = 0.01,
-                                   regularisation_iterations = 300,
-                                   lipschitz_const = lc)
+# adding Huber data fidelity threhsold
+data.update({'huber_threshold' : 4.5})
+# Run FISTA reconstrucion algorithm with regularisation and HUber data
+RecFISTA_Huber_reg = RectoolsIR.FISTA(data, algorithm_params, regularisation_params)
+
+#  adding RING minimisation component (better model for data with rings - different from GH!)
+data.update({'ring_weights_threshold' : 4.5})
+RecFISTA_HuberRing_reg = RectoolsIR.FISTA(data, algorithm_params, regularisation_params)
 
 plt.figure()
-plt.subplot(121)
+plt.subplot(131)
 plt.imshow(RecFISTA_LS_reg, vmin=0, vmax=3, cmap="gray")
 plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
 plt.title('FISTA-OS-LS-TV reconstruction')
-plt.subplot(122)
+plt.subplot(132)
 plt.imshow(RecFISTA_Huber_reg, vmin=0, vmax=3, cmap="gray")
 plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
 plt.title('FISTA-OS-Huber-TV reconstruction')
+plt.subplot(133)
+plt.imshow(RecFISTA_HuberRing_reg, vmin=0, vmax=3, cmap="gray")
+plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
+plt.title('FISTA-OS-HuberRing-TV reconstruction')
 plt.show()
 
 # calculate errors 
-
 Qtools = QualityTools(phantom_2D, RecFISTA_LS_reg)
 RMSE_FISTA_LS_TV = Qtools.rmse()
 Qtools = QualityTools(phantom_2D, RecFISTA_Huber_reg)
 RMSE_FISTA_HUBER_TV = Qtools.rmse()
+Qtools = QualityTools(phantom_2D, RecFISTA_HuberRing_reg)
+RMSE_FISTA_HUBERRING_TV = Qtools.rmse()
 print("RMSE for FISTA-OS-LS-TV reconstruction is {}".format(RMSE_FISTA_LS_TV))
 print("RMSE for FISTA-OS-Huber-TV is {}".format(RMSE_FISTA_HUBER_TV))
-
-#%%
-RectoolsIR = RecToolsIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension (horizontal)
-                    DetectorsDimV = None,  # DetectorsDimV # detector dimension (vertical) for 3D case only
-                    CenterRotOffset = None, # Center of Rotation (CoR) scalar (for 3D case only)
-                    AnglesVec = angles_rad, # array of angles in radians
-                    ObjSize = P, # a scalar to define reconstructed object dimensions
-                    datafidelity='LS', #data fidelity, choose LS, PWLS, GH (wip), Student (wip)
-                    nonnegativity='ENABLE', # enable nonnegativity constraint (set to 'ENABLE')
-                    OS_number = 12, # the number of subsets, NONE/(or > 1) ~ classical / ordered subsets
-                    tolerance = 1e-06, # tolerance to stop outer iterations earlier
-                    device='gpu')
-
-RecFISTA_Huber_RingModel = RectoolsIR.FISTA(noisy_zing_stripe, 
-                                   huber_data_threshold=3.0,
-                                   ring_model_horiz_window = 11,
-                                   iterationsFISTA = 20, 
-                                   regularisation = 'ROF_TV', 
-                                   regularisation_parameter = 0.01,
-                                   regularisation_iterations = 300,
-                                   lipschitz_const = lc)
-
-plt.figure()
-plt.subplot(121)
-plt.imshow(RecFISTA_Huber_reg, vmin=0, vmax=3, cmap="gray")
-plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
-plt.title('FISTA-OS-Huber-TV reconstruction')
-plt.subplot(122)
-plt.imshow(RecFISTA_Huber_RingModel, vmin=0, vmax=3, cmap="gray")
-plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
-plt.title('FISTA-OS-Huber-RingModel-TV reconstruction')
-plt.show()
-
-# calculate errors 
-Qtools = QualityTools(phantom_2D, RecFISTA_Huber_reg)
-RMSE_FISTA_HUBER_TV = Qtools.rmse()
-Qtools = QualityTools(phantom_2D, RecFISTA_Huber_RingModel)
-RMSE_FISTA_HUBER_RING_TV = Qtools.rmse()
-
-print("RMSE for FISTA-OS-Huber-TV is {}".format(RMSE_FISTA_HUBER_TV))
-print("RMSE for FISTA-OS-Huber-RingModel-TV is {}".format(RMSE_FISTA_HUBER_RING_TV))
-#%%
-RectoolsIR = RecToolsIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension (horizontal)
-                    DetectorsDimV = None,  # DetectorsDimV # detector dimension (vertical) for 3D case only
-                    CenterRotOffset = None, # Center of Rotation (CoR) scalar (for 3D case only)
-                    AnglesVec = angles_rad, # array of angles in radians
-                    ObjSize = P, # a scalar to define reconstructed object dimensions
-                    datafidelity='LS', #data fidelity, choose LS, PWLS, GH (wip), Student (wip)
-                    nonnegativity='ENABLE', # enable nonnegativity constraint (set to 'ENABLE')
-                    OS_number = 12, # the number of subsets, NONE/(or > 1) ~ classical / ordered subsets
-                    tolerance = 1e-10, # tolerance to stop outer iterations earlier
-                    device='gpu')
-
-RecFISTA_Sudentst = RectoolsIR.FISTA(noisy_zing_stripe, 
-                                   student_data_threshold = 4.5,
-                                   iterationsFISTA = 150, 
-                                   regularisation = 'ROF_TV', 
-                                   regularisation_parameter = 0.0008,
-                                   regularisation_iterations = 350,
-                                   lipschitz_const = lc)
-
-plt.figure()
-plt.subplot(121)
-plt.imshow(RecFISTA_Huber_reg, vmin=0, vmax=3, cmap="gray")
-plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
-plt.title('FISTA-OS-Huber-TV reconstruction')
-plt.subplot(122)
-plt.imshow(RecFISTA_Sudentst, vmin=0, vmax=3, cmap="gray")
-plt.colorbar(ticks=[0, 0.5, 1], orientation='vertical')
-plt.title('FISTA-OS-Studentst-TV reconstruction')
-plt.show()
-
-# calculate errors 
-Qtools = QualityTools(phantom_2D, RecFISTA_Huber_reg)
-RMSE_FISTA_HUBER_TV = Qtools.rmse()
-Qtools = QualityTools(phantom_2D, RecFISTA_Sudentst)
-RMSE_FISTA_STUDENT_TV = Qtools.rmse()
-
-print("RMSE for FISTA-OS-Huber-TV is {}".format(RMSE_FISTA_HUBER_TV))
-print("RMSE for FISTA-OS-Studentst-TV is {}".format(RMSE_FISTA_STUDENT_TV))
+print("RMSE for FISTA-OS-HuberRing-TV is {}".format(RMSE_FISTA_HUBERRING_TV))
 #%%
 print ("%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%")
 print ("Reconstructing using FISTA-Group-Huber method (tomobar)")
 print ("%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%")
-
 from tomobar.methodsIR import RecToolsIR
-
 RectoolsIR = RecToolsIR(DetectorsDimH = P,  # DetectorsDimH # detector dimension (horizontal)
                     DetectorsDimV = None,  # DetectorsDimV # detector dimension (vertical) for 3D case only
                     CenterRotOffset = None, # Center of Rotation (CoR) scalar (for 3D case only)
                     AnglesVec = angles_rad, # array of angles in radians
                     ObjSize = P, # a scalar to define reconstructed object dimensions
-                    datafidelity='LS', #data fidelity, choose LS, PWLS, Huber, GH (wip), Student (wip)
-                    nonnegativity='ENABLE', # enable nonnegativity constraint (set to 'ENABLE')
+                    datafidelity='LS', #data fidelity, choose LS
                     OS_number = None, # the number of subsets, NONE/(or > 1) ~ classical / ordered subsets
-                    tolerance = 1e-06, # tolerance to stop outer iterations earlier
-                    device='gpu')
+                    device_projector='gpu')
 
-lc = RectoolsIR.powermethod() # calculate Lipschitz constant
+
+# prepare dictionaries with parameters:
+data = {'projection_norm_data' : noisy_zing_stripe,
+        'ringGH_lambda' : 0.0025,
+        'ringGH_accelerate': 100,
+        } # data dictionary
+lc = RectoolsIR.powermethod(data) # calculate Lipschitz constant (run once to initialise)
+algorithm_params = {'iterations' : 350,
+                    'lipschitz_const' : lc}
+
+# adding regularisation using the CCPi regularisation toolkit
+regularisation_params = {'method' : 'FGP_TV',
+                         'regul_param' : 0.001,
+                         'iterations' : 100,
+                         'device_regulariser': 'gpu'}
 
 # Run FISTA reconstrucion algorithm with regularisation 
-RecFISTA_LS_GH_reg = RectoolsIR.FISTA(noisy_zing_stripe, 
-                                   lambdaR_L1 = 0.0025,
-                                   alpha_ring = 150,
-                                   iterationsFISTA = 150, 
-                                   regularisation = 'ROF_TV', 
-                                   regularisation_parameter = 0.01,
-                                   regularisation_iterations = 80,
-                                   lipschitz_const = lc)
+RecFISTA_LS_GH_reg = RectoolsIR.FISTA(data, algorithm_params, regularisation_params)
 
 plt.figure()
 plt.imshow(RecFISTA_LS_GH_reg, vmin=0, vmax=3, cmap="gray")